+++ guillotina/tests/test_swagger.py
@@ -1,3 +1,6 @@
+from openapi_spec_validator import validate_v3_spec
+
+import os
 import pytest
 
 
@@ -30,3 +33,35 @@ async def test_get_swagger_index(container_requester):
     async with container_requester as requester:
         resp, status = await requester("GET", "/@docs")
         assert status == 200
+
+
+@pytest.mark.app_settings(SWAGGER_SETTINGS)
+async def test_validate_swagger_definition(container_requester):
+    async with container_requester as requester:
+        for path in ("/", "/db", "/db/guillotina"):
+            resp, status = await requester("GET", os.path.join(path, "@swagger"))
+            assert status == 200
+            validate_v3_spec(resp)
+
+
+async def test_validate_arrays_in_query_params(container_requester):
+    async with container_requester as requester:
+        _, status = await requester("GET", "@queryParamsValidation?numbers=1.0&numbers=2")
+        assert status == 200
+
+        _, status = await requester("GET", "@queryParamsValidation?users=u1&users=u2&users=u3")
+        assert status == 200
+
+        resp, status = await requester("GET", "@queryParamsValidation?users=u1&users=")
+        assert status == 412
+        assert resp["validator"] == "minLength"
+
+        resp, status = await requester("GET", "@queryParamsValidation?users=u1")
+        assert status == 412
+        assert resp["validator"] == "minItems"
+
+        resp, status = await requester(
+            "GET", "@queryParamsValidation?users=u1&users=u2&users=u3&users=u4&users=u5&users=u6"
+        )
+        assert status == 412
+        assert resp["validator"] == "maxItems"